requirements.txt
cryptography>=41.0.0
paramiko>=2.8.1
bcrypt>=3.2.0
PyJWT>=2.3.0
argon2-cffi>=21.3.0
pynacl>=1.4.0
requests>=2.31.0
aiohttp>=3.9.0
python-dateutil>=2.8.2
typing-extensions>=4.8.0
